# openapi.yml
openapi: 3.0.3
info:
  title: ASPB Autowebinar API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/admin/moderation/reports:
    get:
      summary: List platform moderation reports
      security: [{ adminSessionCookie: [] }]
      parameters:
        - { in: query, name: status, schema: { $ref: '#/components/schemas/ModerationCaseStatus' } }
        - { in: query, name: organizationId, schema: { type: string, maxLength: 191 } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200': { description: Platform queue without reporter identity; owner/admin only }
        '401': { description: MFA AdminUser session is missing or invalid }
        '403': { description: AdminUser role is not owner/admin }
  /api/admin/moderation/reports/{id}/status:
    patch:
      summary: Apply an allowed moderation state transition
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ModerationTransitionRequest' } } }
      responses:
        '200': { description: State, immutable event and audit committed atomically }
        '404': { description: Unknown report }
        '409':
          { description: Invalid transition or optimistic revision conflict; rejected mutation creates no event/audit }
  /api/admin/moderation/reports/{id}/actions:
    post:
      summary: Unpublish/restore a Webinar or suspend/restore an author
      description: Requires MFA platform owner/admin authorization, reason, explicit confirmation and both report/target revisions. Existing evidence, CRM and analytics are preserved.
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ModerationPlatformActionRequest' } } }
      responses:
        '200': { description: Reversible action and audit committed atomically }
        '409': { description: Action rollout disabled, stale revision or restoration evidence unavailable }
  /api/admin/moderation/reports/{id}/correction-requests:
    post:
      summary: Request a reasoned human-reviewed Webinar correction
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CorrectionRequestCommand' } } }
      responses:
        '201': { description: Immutable correction request created; explicit visibility decision applied }
        '409': { description: Report revision conflict }
  /api/admin/moderation/corrections/{id}/review:
    post:
      summary: Approve or reject a submitted Webinar revision after human review
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CorrectionReviewRequest' } } }
      responses:
        '200': { description: Human decision audited; approved content is versioned and only then becomes current }
        '409': { description: Request or Webinar content version changed }
  /api/admin/analytics/organizations:
    get:
      summary: Read platform-admin organization/Webinar/Session aggregates
      description: Returns privacy-thresholded aggregates only. Raw chat, notes, email, phone and Telegram identifiers are structurally excluded; no tenant User role grants access.
      security: [{ adminSessionCookie: [] }]
      parameters:
        - { in: query, name: organizationId, schema: { type: string, maxLength: 191 } }
        - { in: query, name: webinarId, schema: { type: string, maxLength: 191 } }
        - { in: query, name: sessionId, schema: { type: string, maxLength: 191 } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
      responses:
        '200': { description: UTC aggregates with counts below three suppressed }
        '401': { description: MFA AdminUser session missing; tenant sessions are not accepted }
        '403': { description: AdminUser role is not owner/admin }
        '404': { description: Unknown scoped filter }
  /api/admin/platform/feature-flags:
    get:
      summary: List managed fail-closed platform feature flags
      security: [{ adminSessionCookie: [] }]
      responses:
        '200': { description: Safe flag values and optimistic revisions; owner/admin only }
  /api/admin/platform/feature-flags/{key}:
    patch:
      summary: Change a managed fail-closed feature flag
      description: Owner only. Mutation never sends messages or starts provider jobs.
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: key, required: true, schema: { type: string, pattern: '^[a-z][a-z0-9_]{2,63}$' } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlatformFeatureFlagChange' } } }
      responses:
        '200': { description: Before/after evidence and audit committed with the updated revision }
        '403': { description: AdminUser is not owner }
        '409': { description: Optimistic revision conflict }
  /api/admin/platform/organizations/{id}:
    patch:
      summary: Change platform organization state with rollback evidence
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlatformOrganizationChange' } } }
      responses:
        '200': { description: Owner-only confirmed organization change committed and audited }
        '409': { description: Optimistic revision conflict }
  /api/admin/platform/taxonomy/{kind}/{id}:
    patch:
      summary: Change an allowed legal taxonomy entry
      security: [{ adminSessionCookie: [] }]
      parameters:
        - { in: path, name: kind, required: true, schema: { type: string, enum: [practice_area, jurisdiction] } }
        - { in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlatformTaxonomyChange' } } }
      responses:
        '200': { description: Owner-only confirmed taxonomy change committed and audited }
        '409': { description: Optimistic revision conflict }
  /api/admin/platform/changes/{id}/rollback:
    post:
      summary: Roll back one reversible platform configuration change
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedRevision, reason, confirmation]
              properties:
                expectedRevision: { type: integer, minimum: 1 }
                reason: { type: string, minLength: 3, maxLength: 500 }
                confirmation: { type: string, enum: [CONFIRM_PLATFORM_ROLLBACK] }
      responses:
        '200': { description: Current revision restored from immutable before-state and rollback audited }
        '409': { description: Change already rolled back or target has since changed }
...
  /api/admin/platform/tenant-rollouts:
    get:
      summary: List DB-backed tenant rollout policies and tenant entries
      security: [{ adminSessionCookie: [] }]
      responses:
        '200': { description: Private no-store policy revisions and entries }
        '403': { description: Platform owner/admin role and verified MFA required }
  /api/admin/platform/tenant-rollouts/{feature}:
    patch:
      summary: Change a rollout mode with confirmation, reason and optimistic revision
      security: [{ adminSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/RolloutFeature'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RolloutPolicyUpdateRequest' } } }
      responses:
        '200': { description: Policy changed and audited }
        '403': { description: Platform owner role and verified MFA required }
        '409': { description: Policy revision is stale }
  /api/admin/platform/tenant-rollouts/{feature}/organizations/{organizationId}:
    put:
      summary: Set one tenant entry for an allowlist policy
      security: [{ adminSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/RolloutFeature'
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RolloutEntryUpdateRequest' } } }
      responses:
        '200': { description: Tenant entry changed and audited }
        '403': { description: Platform owner role and verified MFA required }
        '409': { description: Entry revision is stale }
  /api/admin/platform/legal-holds:
    post:
      summary: Create a category-scoped legal hold
      security: [{ adminSessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LegalHoldCreateRequest' } } }
      responses:
        '201': { description: Legal hold created and audited }
        '403': { description: Platform owner role and verified MFA required }
  /api/admin/platform/legal-holds/{legalHoldId}/release:
    post:
      summary: Release a legal hold using optimistic revision
      security: [{ adminSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/LegalHoldId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LegalHoldReleaseRequest' } } }
      responses:
        '200': { description: Legal hold released and audited }
        '403': { description: Platform owner role and verified MFA required }
        '404': { description: Legal hold is absent }
        '409': { description: Stale revision or already released }
components:
  parameters:
...
    CsrfToken:
      name: x-csrf-token
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 4096 }
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
...
    LegalHoldId:
      name: legalHoldId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
    RolloutFeature:
      name: feature
      in: path
      required: true
      schema:
        type: string
        enum:
          [
            PLATFORM_ACCOUNTS_ONBOARDING,
            CREATOR_DASHBOARD,
            PUBLIC_CATALOG,
            TENANT_CRM,
            TENANT_TELEGRAM,
            PROVIDER_JOBS,
            ANALYTICS_MODERATION,
          ]
...
  schemas:
...
    RolloutPolicyUpdateRequest:
      type: object
      additionalProperties: false
      required: [mode, expectedRevision, reason, confirm]
      properties:
        mode: { type: string, enum: [DISABLED, ALLOWLIST, ENABLED] }
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 10, maxLength: 1000 }
        confirm: { type: boolean, enum: [true] }
    RolloutEntryUpdateRequest:
      type: object
      additionalProperties: false
      required: [enabled, expectedRevision, reason, confirm]
      properties:
        enabled: { type: boolean }
        expectedRevision: { type: integer, minimum: 0 }
        reason: { type: string, minLength: 10, maxLength: 1000 }
        confirm: { type: boolean, enum: [true] }
    RetentionCategory:
      type: string
      enum:
        [
          AUTHOR_PROFILE_EVIDENCE,
          WEBINAR_ACCESS_GRANTS,
          VIEWER_ACCOUNT_DATA,
          TENANT_CRM_DATA,
          MODERATION_CHAT_QUESTION_HISTORY,
          TENANT_TELEGRAM_DATA,
        ]
    LegalHoldCreateRequest:
      type: object
      additionalProperties: false
      required: [organizationId, categories, reason, startsAt, confirm]
      properties:
        organizationId: { type: string, minLength: 1, maxLength: 191 }
        categories:
          type: array
          minItems: 1
          maxItems: 20
          items:
            oneOf:
              - { type: string, enum: ['*'] }
              - { $ref: '#/components/schemas/RetentionCategory' }
        reason: { type: string, minLength: 20, maxLength: 2000 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, nullable: true }
        confirm: { type: boolean, enum: [true] }
    LegalHoldReleaseRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision, reason, confirm]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 20, maxLength: 2000 }
        confirm: { type: boolean, enum: [true] }
...
    ModerationCaseStatus:
      type: string
      enum: [NEW, IN_REVIEW, ACTION_REQUIRED, RESOLVED, REJECTED]
    ModerationTransitionRequest:
      type: object
      additionalProperties: false
      required: [status, expectedRevision, reason]
      properties:
        status: { $ref: '#/components/schemas/ModerationCaseStatus' }
        expectedRevision: { type: integer, minimum: 0 }
        reason: { type: string, minLength: 3, maxLength: 500 }
    ModerationPlatformActionRequest:
      type: object
      additionalProperties: false
      required: [action, expectedRevision, expectedTargetRevision, reason, confirmation]
      properties:
        action: { type: string, enum: [UNPUBLISH_WEBINAR, SUSPEND_AUTHOR, RESTORE_WEBINAR, RESTORE_AUTHOR] }
        expectedRevision: { type: integer, minimum: 0 }
        expectedTargetRevision: { type: integer, minimum: 0 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        confirmation: { type: string, enum: [APPLY_MODERATION_ACTION] }
    CorrectionRequestCommand:
      type: object
      additionalProperties: false
      required: [expectedRevision, reason, visibilityDecision, confirmation]
      properties:
        expectedRevision: { type: integer, minimum: 0 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        visibilityDecision: { type: string, enum: [KEEP_PUBLISHED, HIDE_UNTIL_APPROVED] }
        confirmation: { type: string, enum: [REQUEST_CORRECTION] }
...
    CorrectionReviewRequest:
      type: object
      additionalProperties: false
      required: [decision, expectedRevision, reason, confirmation]
      properties:
        decision: { type: string, enum: [APPROVE, REJECT] }
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        confirmation: { type: string, enum: [REVIEW_CORRECTION] }
    PlatformFeatureFlagChange:
      type: object
      additionalProperties: false
      required: [enabled, expectedRevision, reason, confirmation]
      properties:
        enabled: { type: boolean }
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        confirmation: { type: string, enum: [CONFIRM_PLATFORM_CHANGE] }
    PlatformOrganizationChange:
      type: object
      additionalProperties: false
      required: [expectedRevision, reason, confirmation]
      properties:
        name: { type: string, minLength: 2, maxLength: 191 }
        status: { type: string, enum: [ACTIVE, SUSPENDED, ARCHIVED] }
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        confirmation: { type: string, enum: [CONFIRM_PLATFORM_CHANGE] }
    PlatformTaxonomyChange:
      type: object
      additionalProperties: false
      required: [expectedRevision, reason, confirmation]
      properties:
        name: { type: string, minLength: 2, maxLength: 191 }
        status: { type: string, enum: [ACTIVE, ARCHIVED] }
        orderIndex: { type: integer, minimum: 0, maximum: 100000 }
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        confirmation: { type: string, enum: [CONFIRM_PLATFORM_CHANGE] }
...
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    userSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_user_session
    participantSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_room_token
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_admin_session